Rename equivalences
