+++ src/paritea/equivalence/__init__.py
@@ -1,5 +1,7 @@
-from .check_fault_equivalence import is_fault_equivalence
+from .fault_equivalence import is_fault_equivalence
+from .semantic_equivalence import is_semantic_equivalence
 
 __all__ = [
     "is_fault_equivalence",
+    "is_semantic_equivalence",
 ]
