+++ src/paritea/equivalence/fault_equivalence.py
